refactor: limital tests,rendered the pages separately
